update(Dockerfile): revert patch release change
During the big incident, we had to come up with the quick solution for building Docker image as fast as possible and for that we had to modify and hardcode the Dockerfile. This Pull Request makes sure things are back to the proper state.

In the long term, we are going to deprecate the K8s version of this Node.js-based Agent but for now let's keep this in this way.

Semver: patch