Don't check for root user when program.config isset
If a user defines a program.config, there is really no need to check the root user. We can assume the user wants to put the config in a folder that is not /etc. no?
I don't see any other reason to have the index.js file check for root.