fix(TailReadStream): Cancel timers to prevent shutdown hanging
There was an issue when running the agent as a service where the timers
in `TailReadStream` were hanging the node process. This might be a
difference between the `nexe` binary version and the plain version
(since this issue is not seen when running via node.js directly).
This change adds tracking for all timers in that code, and issues a
`cancelTimeout` for each of them during shutdown. The problem was
verified/reproduced and fixed using a Docker image of Debian.

NOTE: There is currently no test in the test suite that can verify this
change since it already shuts down properly there. Future versions of
the agent will, however, include proper testing and will not be using
`TailReadStream` for file consumption.

Semver: patch
Ref: LOG-7930
Fixes: https://github.com/logdna/logdna-agent/issues/212